+++ jwst/source_catalog/source_catalog.py
@@ -125,6 +125,22 @@ class JWSTSourceCatalog:
         self.model.meta.bunit_data = unit.name
         self.model.meta.bunit_err = unit.name
 
+    def convert_from_jy(self):
+        """
+        Convert the data and errors from Jy to MJy/sr and change from
+        `~astropy.unit.Quantity` objects to `~numpy.ndarray`.
+        """
+
+        to_mjy_sr = 1.e6 * self.model.meta.photometry.pixelarea_steradians
+        self.model.data /= to_mjy_sr
+        self.model.err /= to_mjy_sr
+
+        self.model.data = self.model.data.value  # remove units
+        self.model.err = self.model.err.value  # remove units
+
+        self.model.meta.bunit_data = 'MJy/sr'
+        self.model.meta.bunit_err = 'MJy/sr'
+
     @staticmethod
     def convert_flux_to_abmag(flux, flux_err):
         """
@@ -947,5 +963,8 @@ class JWSTSourceCatalog:
         self.meta['aperture_params'] = self.aperture_params
         self.meta['abvega_offset'] = self.abvega_offset
         catalog.meta.update(self.meta)
+
+        # reset units on input model back to MJy/sr
+        self.convert_from_jy()
 
         return catalog